解决用户数据权限查询出重复的问题

+++ uap/src/main/java/com/dili/uap/service/impl/UserServiceImpl.java
@@ -397,24 +397,50 @@ public class UserServiceImpl extends BaseServiceImpl<User, Long> implements User
 		if (userTicket == null) {
 			return BaseOutput.failure("用户未登录");
 		}
+		BaseOutput<Object> output = null;
+		List<UserDataAuth> saveDatas = this.convertToSaveUserDatas(userId, dataRange, dataIds);
 		UserDataAuth record = DTOUtils.newInstance(UserDataAuth.class);
 		record.setUserId(userId);
-		List<UserDataAuth> currentUserDataAuthList = this.userDataAuthMapper.select(record);
-		record.setUserId(userTicket.getId());
-		List<UserDataAuth> loggedUserDataAuthList = this.userDataAuthMapper.select(record);
-		if (CollectionUtils.isEmpty(loggedUserDataAuthList)) {
-			return BaseOutput.failure("当前登录用户没有数据权限！");
+		if (this.adminName.equals(userTicket.getUserName())) {
+			this.userDataAuthMapper.delete(record);
+		} else {
+			List<UserDataAuth> currentUserDataAuthList = this.userDataAuthMapper.select(record);
+			record.setUserId(userTicket.getId());
+			List<UserDataAuth> loggedUserDataAuthList = this.userDataAuthMapper.select(record);
+			if (CollectionUtils.isEmpty(loggedUserDataAuthList)) {
+				return BaseOutput.failure("当前登录用户没有数据权限！");
+			}
+			List<Long> toRmoveIds = new ArrayList<>();
+			List<UserDataAuth> toRemoveList = currentUserDataAuthList.stream()
+					.filter(ud -> loggedUserDataAuthList.stream().filter(lud -> lud.getRefCode().equals(ud.getRefCode()) && lud.getValue().equals(ud.getValue())).findFirst().orElse(null) != null)
+					.collect(Collectors.toList());
+			if (CollectionUtils.isNotEmpty(toRemoveList)) {
+				toRemoveList.forEach(ud -> toRmoveIds.add(ud.getId()));
+				Example example = new Example(UserDataAuth.class);
+				example.createCriteria().andEqualTo("userId", userId).andIn("id", toRmoveIds);
+				this.userDataAuthMapper.deleteByExample(example);
+			}
+			int originalSize = saveDatas.size();
+			saveDatas = saveDatas.stream()
+					.filter(sd -> toRemoveList.stream().filter(rud -> rud.getRefCode().equals(sd.getRefCode()) && rud.getValue().equals(sd.getValue())).findFirst().orElse(null) != null)
+					.collect(Collectors.toList());
+			saveDatas = saveDatas.stream()
+					.filter(sd -> loggedUserDataAuthList.stream().filter(lud -> lud.getRefCode().equals(sd.getRefCode()) && lud.getValue().equals(sd.getValue())).findFirst().orElse(null) != null)
+					.collect(Collectors.toList());
+			if (currentUserDataAuthList.size() != toRemoveList.size() || originalSize != saveDatas.size()) {
+				output = BaseOutput.success("修改的用户数据权限中包含当前登录用户不具备的数据权限，部分数据修改不成功，请核对数据！");
+			}
 		}
-		List<Long> toRmoveIds = new ArrayList<>();
-		List<UserDataAuth> toRemoveList = currentUserDataAuthList.stream()
-				.filter(ud -> loggedUserDataAuthList.stream().filter(lud -> lud.getRefCode().equals(ud.getRefCode()) && lud.getValue().equals(ud.getValue())).findFirst().orElse(null) != null)
-				.collect(Collectors.toList());
-		toRemoveList.forEach(ud -> toRmoveIds.add(ud.getId()));
-		Example example = new Example(UserDataAuth.class);
-		example.createCriteria().andEqualTo("userId", userId).andIn("id", toRmoveIds);
-		this.userDataAuthMapper.deleteByExample(example);
-		List<UserDataAuth> saveDatas = Lists.newArrayList();
+		// 如果存在需要保存的用户角色数据，则保存数据
+		if (CollectionUtils.isNotEmpty(saveDatas)) {
+			userDataAuthMapper.insertList(saveDatas);
+		}
+		return output == null ? BaseOutput.success("操作成功") : output;
+	}
+
+	private List<UserDataAuth> convertToSaveUserDatas(Long userId, Long dataRange, String[] dataIds) {
 		// 保存用户数据范围信息
+		List<UserDataAuth> saveDatas = new ArrayList<>();
 		UserDataAuth ud = DTOUtils.newInstance(UserDataAuth.class);
 		ud.setUserId(userId);
 		ud.setValue(String.valueOf(dataRange));
@@ -444,21 +470,7 @@ public class UserServiceImpl extends BaseServiceImpl<User, Long> implements User
 				}
 			}
 		}
-		saveDatas = saveDatas.stream()
-				.filter(sd -> toRemoveList.stream().filter(rud -> rud.getRefCode().equals(sd.getRefCode()) && rud.getValue().equals(sd.getValue())).findFirst().orElse(null) != null)
-				.collect(Collectors.toList());
-		int originalSize = saveDatas.size();
-		saveDatas = saveDatas.stream()
-				.filter(sd -> loggedUserDataAuthList.stream().filter(lud -> lud.getRefCode().equals(sd.getRefCode()) && lud.getValue().equals(sd.getValue())).findFirst().orElse(null) != null)
-				.collect(Collectors.toList());
-		// 如果存在需要保存的用户角色数据，则保存数据
-		if (CollectionUtils.isNotEmpty(saveDatas)) {
-			userDataAuthMapper.insertList(saveDatas);
-		}
-		if (currentUserDataAuthList.size() != toRemoveList.size() || originalSize != saveDatas.size()) {
-			return BaseOutput.success("修改的用户数据权限中包含当前登录用户不具备的数据权限，部分数据修改不成功，请核对数据！");
-		}
-		return BaseOutput.success("操作成功");
+		return saveDatas;
 	}
 
 	@Override
@@ -556,7 +568,6 @@ public class UserServiceImpl extends BaseServiceImpl<User, Long> implements User
 				// 判断选中
 				for (UserDataDto userDataDto : selectAll) {
 					String replace = userDataDto.getTreeId().replace(ALM_PROJECT_PREFIX, "");
-					System.out.println(replace);
 					boolean isChecked = selectUserDataAuthValue.contains(replace);
 					if (!isRootChecked && isChecked) {
 						isRootChecked = true;
